+++ src/Logger.h
@@ -11,5 +11,6 @@ enum class LogLevel {
 
 extern LogLevel g_logLevel;
 
+void InitLogger();
 void Log(const std::wstring& message, LogLevel level = LogLevel::LOG_INFO);
 void CloseLogFile();